extracted pirate class
