+++ code/whack-a-pirate.py
@@ -6,6 +6,7 @@ import numpy as np
 #import psychopy
 from math import sin, cos, pi
 from Button import *
+from Pirate import *
 
 # Initialize Pygame
 pygame.init()
@@ -81,34 +82,6 @@ print('Actual flicker frequencies = ', sorted(actual_freqs))
 # Calculate the cycle durations for each frequency (1/frequency)
 # Calculate the cycle durations for each frequency (1/frequency)
 
-
-# Define Pirate class
-class Pirate(pygame.sprite.Sprite):
-    def __init__(self, image, sil, skull_image, location, duration):
-        super().__init__()
-        self.image = image
-        self.sil = sil
-        self.skull_image = skull_image
-        self.rect = self.image.get_rect()
-        self.duration = duration
-        self.location = location
-        self.visible = False
-        self.clicked = False
-
-    def update(self):
-        if self.location is not None:
-            self.rect.x = self.location[0] - self.rect.width / 2
-            self.rect.y = self.location[1] - self.rect.height / 2
-
-    def draw(self, surface):
-        if self.visible:
-            if self.clicked:
-                surface.blit(self.skull_image, self.rect)
-            else:
-                surface.blit(self.image, self.rect)
-
-    def draw_silhouette(self, surface):
-            surface.blit(self.sil, self.rect)
 
 # Calculate the cycle durations for each frequency (1/frequency)
 durations = [1.0/freq for freq in actual_freqs]
